fix: request status messages explicitly
This works around a bug with the current driver implementation. Namely, the ping
protocol is throwing away any packets that might come between the command and
the reply. This includes the status message, which explains why we start losing
status messages if the ping frequency is higher.

In addition, it restricts the status data to what we actually use, which reduces
the load on the beacon (something the documentation warns about)